% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aside.R
\name{aside}
\alias{aside}
\title{Aside}
\usage{
aside(cmd, show = FALSE, beep = FALSE)
}
\arguments{
\item{cmd}{the cmd to send}

\item{show}{whether or not to show the terminal}

\item{beep}{whether or not to play a sound when the job is done, using the beepr package when available.}
}
\value{
a job well done
}
\description{
Send the job aside
}
