% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aside.R
\name{aside}
\alias{aside}
\title{Aside}
\usage{
aside(cmd, show = FALSE)
}
\arguments{
\item{cmd}{the cmd to send}

\item{show}{wether or not to show the terminal}
}
\value{
a job well done
}
\description{
Send the job aside
}
